split books and movies tools into two files
