+++ src/agents/books_agent.py
@@ -2,7 +2,7 @@ from pathlib import Path
 import sys
 
 sys.path.append(str(Path(__file__).parent.parent.parent))
-from src.tools.books_and_movies_tools import (
+from src.tools.books_tools import (
     add_book_to_reading_list,
     get_book_lists,
     mark_book_read,
